+++ notify-impl/src/main/java/com/dwarfeng/notify/impl/handler/NotifyHandlerImpl.java
@@ -2,16 +2,12 @@ package com.dwarfeng.notify.impl.handler;
 
 import com.dwarfeng.notify.stack.bean.dto.NotifyInfo;
 import com.dwarfeng.notify.stack.bean.dto.Routing;
-import com.dwarfeng.notify.stack.bean.entity.RouterInfo;
-import com.dwarfeng.notify.stack.bean.entity.SenderInfo;
-import com.dwarfeng.notify.stack.bean.entity.SenderRelation;
+import com.dwarfeng.notify.stack.bean.entity.*;
 import com.dwarfeng.notify.stack.bean.entity.key.SenderRelationKey;
 import com.dwarfeng.notify.stack.exception.RouterException;
 import com.dwarfeng.notify.stack.exception.SenderException;
 import com.dwarfeng.notify.stack.handler.*;
-import com.dwarfeng.notify.stack.service.RouterInfoMaintainService;
-import com.dwarfeng.notify.stack.service.SenderInfoMaintainService;
-import com.dwarfeng.notify.stack.service.SenderRelationMaintainService;
+import com.dwarfeng.notify.stack.service.*;
 import com.dwarfeng.subgrade.stack.bean.key.LongIdKey;
 import com.dwarfeng.subgrade.stack.bean.key.StringIdKey;
 import com.dwarfeng.subgrade.stack.exception.HandlerException;
@@ -20,34 +16,48 @@ import org.slf4j.LoggerFactory;
 import org.springframework.stereotype.Component;
 
 import java.util.*;
+import java.util.stream.Collectors;
 
 @Component
 public class NotifyHandlerImpl implements NotifyHandler {
 
     private static final Logger LOGGER = LoggerFactory.getLogger(NotifyHandlerImpl.class);
 
+    private final NotifySettingMaintainService notifySettingMaintainService;
     private final RouterInfoMaintainService routerInfoMaintainService;
     private final SenderInfoMaintainService senderInfoMaintainService;
     private final SenderRelationMaintainService senderRelationMaintainService;
+    private final TopicMaintainService topicMaintainService;
+    private final UserMaintainService userMaintainService;
 
     private final RouteLocalCacheHandler routeLocalCacheHandler;
     private final SendLocalCacheHandler sendLocalCacheHandler;
 
+    private final PushHandler pushHandler;
+
     private final HandlerValidator handlerValidator;
 
     public NotifyHandlerImpl(
+            NotifySettingMaintainService notifySettingMaintainService,
             RouterInfoMaintainService routerInfoMaintainService,
             SenderInfoMaintainService senderInfoMaintainService,
             SenderRelationMaintainService senderRelationMaintainService,
+            TopicMaintainService topicMaintainService,
+            UserMaintainService userMaintainService,
             RouteLocalCacheHandler routeLocalCacheHandler,
             SendLocalCacheHandler sendLocalCacheHandler,
+            PushHandler pushHandler,
             HandlerValidator handlerValidator
     ) {
+        this.notifySettingMaintainService = notifySettingMaintainService;
         this.routerInfoMaintainService = routerInfoMaintainService;
         this.senderInfoMaintainService = senderInfoMaintainService;
         this.senderRelationMaintainService = senderRelationMaintainService;
+        this.topicMaintainService = topicMaintainService;
+        this.userMaintainService = userMaintainService;
         this.routeLocalCacheHandler = routeLocalCacheHandler;
         this.sendLocalCacheHandler = sendLocalCacheHandler;
+        this.pushHandler = pushHandler;
         this.handlerValidator = handlerValidator;
     }
 
@@ -59,6 +69,7 @@ public class NotifyHandlerImpl implements NotifyHandler {
 
             // 确认 notifySettingKey 存在。
             handlerValidator.makeSureNotifySettingExists(notifySettingKey);
+            NotifySetting notifySetting = notifySettingMaintainService.get(notifySettingKey);
 
             // 查找 notifySettingKey 对应的所有路由器信息，并通过本地缓存拿出路由器。
             List<RouterInfo> routerInfos = routerInfoMaintainService.lookupAsList(
@@ -99,11 +110,35 @@ public class NotifyHandlerImpl implements NotifyHandler {
                 }
             }
 
+            // 临时定义本地映射，这些映射缓存了主键与实体的对应关系，避免后续代码频繁操作数据访问层进行读取。
+            Map<StringIdKey, Topic> topicMap = new HashMap<>();
+            Map<StringIdKey, User> userMap = new HashMap<>();
+
+            // 遍历映射键，确认主题全部存在，查出实体并存放在本地映射中。
+            for (StringIdKey topicKey : routingMap.keySet()) {
+                handlerValidator.makeSureTopicExists(topicKey);
+                Topic topic = topicMaintainService.get(topicKey);
+                topicMap.put(topicKey, topic);
+            }
+            // 遍历映射值，确认用户全部存在，查出实体并存放在本地映射中。
+            for (List<StringIdKey> userKeys : routingMap.values()) {
+                for (StringIdKey userKey : userKeys) {
+                    if (userMap.containsKey(userKey)) {
+                        continue;
+                    }
+                    handlerValidator.makeSureUserExists(userKey);
+                    User user = userMaintainService.get(userKey);
+                    userMap.put(userKey, user);
+                }
+            }
+
             // 遍历映射入口，发送数据。
             for (Map.Entry<StringIdKey, List<StringIdKey>> entry : routingMap.entrySet()) {
                 StringIdKey topicKey = entry.getKey();
                 List<StringIdKey> userKeys = entry.getValue();
-                notifySingleTopic(notifySettingKey, topicKey, userKeys, context);
+                Topic topic = topicMap.get(topicKey);
+                List<User> users = userKeys.stream().map(userMap::get).collect(Collectors.toList());
+                notifySingleTopic(notifySettingKey, topicKey, userKeys, context, notifySetting, topic, users);
             }
         } catch (HandlerException e) {
             throw e;
@@ -113,7 +148,8 @@ public class NotifyHandlerImpl implements NotifyHandler {
     }
 
     private void notifySingleTopic(
-            LongIdKey notifySettingKey, StringIdKey topicKey, List<StringIdKey> userKeys, Object context
+            LongIdKey notifySettingKey, StringIdKey topicKey, List<StringIdKey> userKeys, Object context,
+            NotifySetting notifySetting, Topic topic, List<User> users
     ) throws Exception {
         // 根据通知设置和主题找到对应的发送器，如果找不到，则警告并退出方法。
         SenderRelationKey senderRelationKey = new SenderRelationKey(
@@ -139,6 +175,7 @@ public class NotifyHandlerImpl implements NotifyHandler {
         // 调用发送器的批量发送方法。
         try {
             sender.batchSend(userKeys, context);
+            pushHandler.notifyHappened(notifySetting, topic, users, context);
         } catch (SenderException e) {
             LOGGER.warn("发送器 " + sender + " 发送信息失败，异常信息如下: ", e);
         }