+++ notify-impl/src/main/java/com/dwarfeng/notify/impl/handler/pusher/LogPusher.java
@@ -1,0 +1,83 @@
+package com.dwarfeng.notify.impl.handler.pusher;
+
+import com.dwarfeng.notify.stack.bean.entity.NotifySetting;
+import com.dwarfeng.notify.stack.bean.entity.Topic;
+import com.dwarfeng.notify.stack.bean.entity.User;
+import com.dwarfeng.subgrade.stack.exception.HandlerException;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.stereotype.Component;
+
+/**
+ * 将信息输出至日志的推送器。
+ *
+ * @author DwArFeng
+ * @since 1.0.0
+ */
+@Component
+public class LogPusher extends AbstractPusher {
+
+    public static final String PUSHER_TYPE = "log";
+
+    private static final Logger LOGGER = LoggerFactory.getLogger(LogPusher.class);
+
+    private static final String LEVEL_TRACE = "TRACE";
+    private static final String LEVEL_DEBUG = "DEBUG";
+    private static final String LEVEL_INFO = "INFO";
+    private static final String LEVEL_WARN = "WARN";
+    private static final String LEVEL_ERROR = "ERROR";
+
+    @Value("${pusher.log.log_level}")
+    private String logLevel;
+
+    public LogPusher() {
+        super(PUSHER_TYPE);
+    }
+
+    @Override
+    public void notifyHappened(NotifySetting notifySetting, Topic topic, User user, Object context)
+            throws HandlerException {
+        String title = "通知发生事件:";
+        String message = String.format(
+                "通知设置: %s, 主题: %s, 用户: %s, 上下文对象: %s", notifySetting, topic, user, context
+        );
+        logData(title, message);
+    }
+
+    private void logData(String title, String message) throws HandlerException {
+        String logLevel = this.logLevel.toUpperCase();
+        switch (logLevel) {
+            case LEVEL_TRACE:
+                LOGGER.trace(title);
+                LOGGER.trace(message);
+                return;
+            case LEVEL_DEBUG:
+                LOGGER.debug(title);
+                LOGGER.debug(message);
+                return;
+            case LEVEL_INFO:
+                LOGGER.info(title);
+                LOGGER.info(message);
+                return;
+            case LEVEL_WARN:
+                LOGGER.warn(title);
+                LOGGER.warn(message);
+                return;
+            case LEVEL_ERROR:
+                LOGGER.error(title);
+                LOGGER.error(message);
+                return;
+            default:
+                throw new HandlerException("未知的日志等级: " + logLevel);
+        }
+    }
+
+    @Override
+    public String toString() {
+        return "LogPusher{" +
+                "pusherType='" + pusherType + '\'' +
+                ", logLevel='" + logLevel + '\'' +
+                '}';
+    }
+}